refactor: Rename to studio
