+++ apps/web/src/components/nav/MainNav.tsx
@@ -1,6 +1,6 @@
 import { ROUTES } from '@novu/shared-web';
 import { FC } from 'react';
-import { LocalNavMenu } from '../../local-studio/components/LocalNavMenu';
+import { LocalNavMenu } from '../../studio/components/LocalNavMenu';
 import { RootNavMenu } from './RootNavMenu';
 import { SettingsNavMenu } from './SettingsNavMenu';
 import { SidebarNav } from './SidebarNav';
